More usage of static_cast to surpress issues

+++ src/lib/sup/dto/json/json_writer_t.h
@@ -111,7 +111,7 @@ bool JSONStringWriterT<WriterImpl>::Bool(boolean b)
 template <typename WriterImpl>
 bool JSONStringWriterT<WriterImpl>::Char(char8 c)
 {
-  return m_json_writer.Int(c);
+  return m_json_writer.Int(static_cast<int>(c));
 }
 
 template <typename WriterImpl>
